addes stocks bought event,listener,email, emails sent to mailtrap

+++ database/migrations/2021_11_21_195259_update_users_table_add_money.php
@@ -14,7 +14,7 @@ class UpdateUsersTableAddMoney extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->bigInteger('money')->default(0);
+            $table->float('money')->default(0);
         });
     }
